requirements.txt
asgiref==3.4.1
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
flake8==4.0.1
importlib-metadata==4.2.0
Markdown==3.3.6
mccabe==0.6.1
Pillow==8.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.0.1
zipp==3.6.0
